+++ src/test/java/com/ghm/giftcardfleamarket/domain/user/service/SessionLoginServiceTest.java
@@ -1,4 +1,4 @@
-package com.ghm.giftcardfleamarket.user.service;
+package com.ghm.giftcardfleamarket.domain.user.service;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
@@ -15,7 +15,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.mock.web.MockHttpSession;
 
 import com.ghm.giftcardfleamarket.domain.user.domain.User;
-import com.ghm.giftcardfleamarket.domain.user.service.SessionLoginService;
 import com.ghm.giftcardfleamarket.global.util.PasswordEncryptor;
 import com.ghm.giftcardfleamarket.global.util.constants.SessionNames;
